python/requirements.txt
blinker==1.9.0
click==8.1.8
crate==1.0.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
urllib3==2.3.0
verlib2==0.2.0
Werkzeug==3.1.3
